Add operating system case sensitivity to README
Database and table names are not case sensitive in Windows, and case
sensitive in most varieties of Unix.

One can configure how tables names are stored on the disk using the
system variable `lower_case_table_names`.

This change provides details in the README file of how to configure the
database to ignore case sensitivity on table names by modifying the
`/etc/mysql/my.cnf` file.

See this post for more details:
- https://stackoverflow.com/questions/6134006/are-table-names-in-mysql-case-sensitive